+++ src/fhops/optimization/heuristics/__init__.py
@@ -1,5 +1,6 @@
 """Heuristic solvers for FHOPS."""
 
+from .multistart import MultiStartResult, run_multi_start
 from .registry import MoveOperator, OperatorContext, OperatorRegistry, SwapOperator
 from .sa import Schedule, solve_sa
 
@@ -10,4 +11,6 @@ __all__ = [
     "OperatorRegistry",
     "SwapOperator",
     "MoveOperator",
+    "run_multi_start",
+    "MultiStartResult",
 ]